backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
numpy>=1.24.0
safetensors>=0.3.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pydub>=0.25.1
ftfy>=6.1.0
regex>=2023.06.03
tqdm>=4.65.0
scikit-learn>=1.3.0
openai-clip>=1.0.0